230214
Changed NCP enum names to be consistent.
Implemented and tested handleMixedExplosive.
Implemented goal booleans to set fred destroyed / omar destroyed for endings.
Added a system color for the quit dialogue.
added lockeddoor booleans
Added to the switchlocation method a way to figure out if doors are locked and not let the player through